backend_django/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-decouple
dj-database-url
Pillow
whitenoise
gunicorn
psycopg2-binary
django-prometheus
